Update link state example to Query components
