+++ examples/todo/package.json
@@ -9,10 +9,9 @@
     "apollo-link-state": "^0.4.1",
     "graphql": "^0.11.7",
     "graphql-tag": "^2.5.0",
-    "lodash.merge": "4.6.0",
     "prop-types": "15.6.0",
     "react": "15.6.2",
-    "react-apollo": "^2.0.0",
+    "react-apollo": "2.1.0-rc.5",
     "react-dom": "15.6.2"
   },
   "devDependencies": {
