+++ examples/todo/src/components/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import Footer from './Footer';
-import AddTodo from '../containers/AddTodo';
-import VisibleTodoList from '../containers/VisibleTodoList';
+import TodoForm from './TodoForm';
+import TodoList from './TodoList';
 
 const App = () => (
   <div>
-    <AddTodo />
-    <VisibleTodoList />
+    <TodoForm />
+    <TodoList />
     <Footer />
   </div>
 );
