+++ examples/todo/src/components/Footer.js
@@ -1,5 +1,27 @@
 import React from 'react';
-import FilterLink from '../containers/FilterLink';
+import { Query } from 'react-apollo';
+import gql from 'graphql-tag';
+
+import Link from './Link';
+
+const GET_VISIBILITY_FILTER = gql`
+  {
+    visibilityFilter @client
+  }
+`;
+
+const FilterLink = ({ filter, children }) => (
+  <Query query={GET_VISIBILITY_FILTER}>
+    {({ data, client }) => (
+      <Link
+        onClick={() => client.writeData({ data: { visibilityFilter: filter } })}
+        active={data.visibilityFilter === filter}
+      >
+        {children}
+      </Link>
+    )}
+  </Query>
+);
 
 const Footer = () => (
   <p>
