+++ examples/todo/src/components/Todo.js
@@ -1,14 +1,26 @@
 import React from 'react';
+import { Mutation } from 'react-apollo';
+import gql from 'graphql-tag';
 
-const Todo = ({ onClick, completed, text }) => (
-  <li
-    onClick={onClick}
-    style={{
-      textDecoration: completed ? 'line-through' : 'none',
-    }}
-  >
-    {text}
-  </li>
+const TOGGLE_TODO = gql`
+  mutation ToggleTodo($id: Int!) {
+    toggleTodo(id: $id) @client
+  }
+`;
+
+const Todo = ({ id, completed, text }) => (
+  <Mutation mutation={TOGGLE_TODO} variables={{ id }}>
+    {toggleTodo => (
+      <li
+        onClick={toggleTodo}
+        style={{
+          textDecoration: completed ? 'line-through' : 'none',
+        }}
+      >
+        {text}
+      </li>
+    )}
+  </Mutation>
 );
 
 export default Todo;
