+++ examples/todo/src/components/TodoList.js
@@ -1,14 +1,43 @@
 import React from 'react';
+import { Query } from 'react-apollo';
+import gql from 'graphql-tag';
+
 import Todo from './Todo';
 
-const TodoList = ({ todos, onTodoClick }) => {
-  return (
-    <ul>
-      {todos.map(todo => (
-        <Todo key={todo.id} {...todo} onClick={() => onTodoClick(todo.id)} />
-      ))}
-    </ul>
-  );
+const GET_TODOS = gql`
+  {
+    todos @client {
+      id
+      completed
+      text
+    }
+    visibilityFilter @client
+  }
+`;
+
+const getVisibleTodos = (todos, filter) => {
+  switch (filter) {
+    case 'SHOW_ALL':
+      return todos;
+    case 'SHOW_COMPLETED':
+      return todos.filter(t => t.completed);
+    case 'SHOW_ACTIVE':
+      return todos.filter(t => !t.completed);
+    default:
+      throw new Error('Unknown filter: ' + filter);
+  }
 };
+
+const TodoList = () => (
+  <Query query={GET_TODOS}>
+    {({ data: { todos, visibilityFilter } }) => (
+      <ul>
+        {getVisibleTodos(todos, visibilityFilter).map(todo => (
+          <Todo key={todo.id} {...todo} />
+        ))}
+      </ul>
+    )}
+  </Query>
+);
 
 export default TodoList;
