--- examples/todo/src/resolvers/visibilityFilter.js
@@ -1,15 +1,0 @@
-const visibilityFilter = {
-  defaults: {
-    visibilityFilter: 'SHOW_ALL',
-  },
-  resolvers: {
-    Mutation: {
-      visibilityFilter: (_, { filter }, { cache }) => {
-        cache.writeData({ data: { visibilityFilter: filter } });
-        return null;
-      },
-    },
-  },
-};
-
-export default visibilityFilter;
